feat: render item page on server side
